wired through Utf8Span and Utf8Memory
