+++ Microsoft.Azure.Cosmos/src/CosmosElements/CosmosString.cs
@@ -4,6 +4,7 @@
 namespace Microsoft.Azure.Cosmos.CosmosElements
 {
     using System;
+    using Microsoft.Azure.Cosmos.Core.Utf8;
     using Microsoft.Azure.Cosmos.Json;
 
 #if INTERNAL
@@ -23,7 +24,7 @@ namespace Microsoft.Azure.Cosmos.CosmosElements
 
         public abstract string Value { get; }
 
-        public abstract bool TryGetBufferedUtf8Value(out ReadOnlyMemory<byte> bufferedUtf8Value);
+        public abstract bool TryGetBufferedValue(out Utf8Memory bufferedUtf8Value);
 
         public override void Accept(ICosmosElementVisitor cosmosElementVisitor)
         {
