+++ Microsoft.Azure.Cosmos/src/Json/JsonBinaryEncoding.SystemStrings.cs
@@ -6,7 +6,7 @@ namespace Microsoft.Azure.Cosmos.Json
     using System;
     using System.Linq;
     using System.Text;
-    using Microsoft.Azure.Cosmos.Core.Collections;
+    using Microsoft.Azure.Cosmos.Core.Utf8;
 
     internal static partial class JsonBinaryEncoding
     {
@@ -51,94 +51,94 @@ namespace Microsoft.Azure.Cosmos.Json
                 "_id",
             };
 
-            public static readonly ReadOnlyMemory<byte>[] Utf8Values = Utf16Values.Select(x => (ReadOnlyMemory<byte>)Encoding.UTF8.GetBytes(x)).ToArray();
+            public static readonly UtfAllString[] UtfAllStringValues = Utf16Values.Select(x => UtfAllString.Create(x)).ToArray();
 
             public static class TokenBuffers
             {
-                public static readonly ReadOnlyMemory<byte> String0 = Encoding.UTF8.GetBytes("$s");
-                public static readonly ReadOnlyMemory<byte> String1 = Encoding.UTF8.GetBytes("$t");
-                public static readonly ReadOnlyMemory<byte> String2 = Encoding.UTF8.GetBytes("$v");
-                public static readonly ReadOnlyMemory<byte> String3 = Encoding.UTF8.GetBytes("_attachments");
-                public static readonly ReadOnlyMemory<byte> String4 = Encoding.UTF8.GetBytes("_etag");
-                public static readonly ReadOnlyMemory<byte> String5 = Encoding.UTF8.GetBytes("_rid");
-                public static readonly ReadOnlyMemory<byte> String6 = Encoding.UTF8.GetBytes("_self");
-                public static readonly ReadOnlyMemory<byte> String7 = Encoding.UTF8.GetBytes("_ts");
-                public static readonly ReadOnlyMemory<byte> String8 = Encoding.UTF8.GetBytes("attachments/");
-                public static readonly ReadOnlyMemory<byte> String9 = Encoding.UTF8.GetBytes("coordinates");
-                public static readonly ReadOnlyMemory<byte> String10 = Encoding.UTF8.GetBytes("geometry");
-                public static readonly ReadOnlyMemory<byte> String11 = Encoding.UTF8.GetBytes("GeometryCollection");
-                public static readonly ReadOnlyMemory<byte> String12 = Encoding.UTF8.GetBytes("id");
-                public static readonly ReadOnlyMemory<byte> String13 = Encoding.UTF8.GetBytes("inE");
-                public static readonly ReadOnlyMemory<byte> String14 = Encoding.UTF8.GetBytes("inV");
-                public static readonly ReadOnlyMemory<byte> String15 = Encoding.UTF8.GetBytes("label");
-                public static readonly ReadOnlyMemory<byte> String16 = Encoding.UTF8.GetBytes("LineString");
-                public static readonly ReadOnlyMemory<byte> String17 = Encoding.UTF8.GetBytes("link");
-                public static readonly ReadOnlyMemory<byte> String18 = Encoding.UTF8.GetBytes("MultiLineString");
-                public static readonly ReadOnlyMemory<byte> String19 = Encoding.UTF8.GetBytes("MultiPoint");
-                public static readonly ReadOnlyMemory<byte> String20 = Encoding.UTF8.GetBytes("MultiPolygon");
-                public static readonly ReadOnlyMemory<byte> String21 = Encoding.UTF8.GetBytes("name");
-                public static readonly ReadOnlyMemory<byte> String22 = Encoding.UTF8.GetBytes("outE");
-                public static readonly ReadOnlyMemory<byte> String23 = Encoding.UTF8.GetBytes("outV");
-                public static readonly ReadOnlyMemory<byte> String24 = Encoding.UTF8.GetBytes("Point");
-                public static readonly ReadOnlyMemory<byte> String25 = Encoding.UTF8.GetBytes("Polygon");
-                public static readonly ReadOnlyMemory<byte> String26 = Encoding.UTF8.GetBytes("properties");
-                public static readonly ReadOnlyMemory<byte> String27 = Encoding.UTF8.GetBytes("type");
-                public static readonly ReadOnlyMemory<byte> String28 = Encoding.UTF8.GetBytes("value");
-                public static readonly ReadOnlyMemory<byte> String29 = Encoding.UTF8.GetBytes("Feature");
-                public static readonly ReadOnlyMemory<byte> String30 = Encoding.UTF8.GetBytes("FeatureCollection");
-                public static readonly ReadOnlyMemory<byte> String31 = Encoding.UTF8.GetBytes("_id");
+                public static readonly Utf8String String0 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("$s"));
+                public static readonly Utf8String String1 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("$t"));
+                public static readonly Utf8String String2 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("$v"));
+                public static readonly Utf8String String3 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("_attachments"));
+                public static readonly Utf8String String4 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("_etag"));
+                public static readonly Utf8String String5 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("_rid"));
+                public static readonly Utf8String String6 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("_self"));
+                public static readonly Utf8String String7 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("_ts"));
+                public static readonly Utf8String String8 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("attachments/"));
+                public static readonly Utf8String String9 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("coordinates"));
+                public static readonly Utf8String String10 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("geometry"));
+                public static readonly Utf8String String11 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("GeometryCollection"));
+                public static readonly Utf8String String12 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("id"));
+                public static readonly Utf8String String13 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("inE"));
+                public static readonly Utf8String String14 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("inV"));
+                public static readonly Utf8String String15 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("label"));
+                public static readonly Utf8String String16 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("LineString"));
+                public static readonly Utf8String String17 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("link"));
+                public static readonly Utf8String String18 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("MultiLineString"));
+                public static readonly Utf8String String19 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("MultiPoint"));
+                public static readonly Utf8String String20 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("MultiPolygon"));
+                public static readonly Utf8String String21 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("name"));
+                public static readonly Utf8String String22 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("outE"));
+                public static readonly Utf8String String23 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("outV"));
+                public static readonly Utf8String String24 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("Point"));
+                public static readonly Utf8String String25 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("Polygon"));
+                public static readonly Utf8String String26 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("properties"));
+                public static readonly Utf8String String27 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("type"));
+                public static readonly Utf8String String28 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("value"));
+                public static readonly Utf8String String29 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("Feature"));
+                public static readonly Utf8String String30 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("FeatureCollection"));
+                public static readonly Utf8String String31 = Utf8String.UnsafeFromUtf8BytesNoValidation(Encoding.UTF8.GetBytes("_id"));
             }
 
-            public static int? GetSystemStringId(ReadOnlySpan<byte> buffer)
+            public static int? GetSystemStringId(Utf8Span buffer)
             {
                 switch (buffer.Length)
                 {
                     case 2:
-                        return GetSystemStringIdLength2(buffer);
+                        return GetSystemStringIdLength2(buffer.Span);
                     case 12:
-                        return GetSystemStringIdLength12(buffer);
+                        return GetSystemStringIdLength12(buffer.Span);
                     case 5:
-                        return GetSystemStringIdLength5(buffer);
+                        return GetSystemStringIdLength5(buffer.Span);
                     case 4:
-                        return GetSystemStringIdLength4(buffer);
+                        return GetSystemStringIdLength4(buffer.Span);
                     case 3:
-                        return GetSystemStringIdLength3(buffer);
+                        return GetSystemStringIdLength3(buffer.Span);
                     case 11:
-                        return GetSystemStringIdLength11(buffer);
+                        return GetSystemStringIdLength11(buffer.Span);
                     case 8:
-                        return GetSystemStringIdLength8(buffer);
+                        return GetSystemStringIdLength8(buffer.Span);
                     case 18:
-                        return GetSystemStringIdLength18(buffer);
+                        return GetSystemStringIdLength18(buffer.Span);
                     case 10:
-                        return GetSystemStringIdLength10(buffer);
+                        return GetSystemStringIdLength10(buffer.Span);
                     case 15:
-                        return GetSystemStringIdLength15(buffer);
+                        return GetSystemStringIdLength15(buffer.Span);
                     case 7:
-                        return GetSystemStringIdLength7(buffer);
+                        return GetSystemStringIdLength7(buffer.Span);
                     case 17:
-                        return GetSystemStringIdLength17(buffer);
+                        return GetSystemStringIdLength17(buffer.Span);
                 }
                 return null;
             }
 
             private static int? GetSystemStringIdLength2(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String0.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String0.Span.Span))
                 {
                     return 0;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String1.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String1.Span.Span))
                 {
                     return 1;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String2.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String2.Span.Span))
                 {
                     return 2;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String12.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String12.Span.Span))
                 {
                     return 12;
                 }
@@ -148,17 +148,17 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength12(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String3.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String3.Span.Span))
                 {
                     return 3;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String8.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String8.Span.Span))
                 {
                     return 8;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String20.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String20.Span.Span))
                 {
                     return 20;
                 }
@@ -168,27 +168,27 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength5(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String4.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String4.Span.Span))
                 {
                     return 4;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String6.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String6.Span.Span))
                 {
                     return 6;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String15.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String15.Span.Span))
                 {
                     return 15;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String24.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String24.Span.Span))
                 {
                     return 24;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String28.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String28.Span.Span))
                 {
                     return 28;
                 }
@@ -198,32 +198,32 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength4(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String5.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String5.Span.Span))
                 {
                     return 5;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String17.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String17.Span.Span))
                 {
                     return 17;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String21.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String21.Span.Span))
                 {
                     return 21;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String22.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String22.Span.Span))
                 {
                     return 22;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String23.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String23.Span.Span))
                 {
                     return 23;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String27.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String27.Span.Span))
                 {
                     return 27;
                 }
@@ -233,22 +233,22 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength3(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String7.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String7.Span.Span))
                 {
                     return 7;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String13.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String13.Span.Span))
                 {
                     return 13;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String14.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String14.Span.Span))
                 {
                     return 14;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String31.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String31.Span.Span))
                 {
                     return 31;
                 }
@@ -258,7 +258,7 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength11(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String9.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String9.Span.Span))
                 {
                     return 9;
                 }
@@ -268,7 +268,7 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength8(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String10.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String10.Span.Span))
                 {
                     return 10;
                 }
@@ -278,7 +278,7 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength18(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String11.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String11.Span.Span))
                 {
                     return 11;
                 }
@@ -288,17 +288,17 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength10(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String16.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String16.Span.Span))
                 {
                     return 16;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String19.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String19.Span.Span))
                 {
                     return 19;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String26.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String26.Span.Span))
                 {
                     return 26;
                 }
@@ -308,7 +308,7 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength15(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String18.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String18.Span.Span))
                 {
                     return 18;
                 }
@@ -318,12 +318,12 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength7(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String25.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String25.Span.Span))
                 {
                     return 25;
                 }
 
-                if (buffer.SequenceEqual(TokenBuffers.String29.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String29.Span.Span))
                 {
                     return 29;
                 }
@@ -333,7 +333,7 @@ namespace Microsoft.Azure.Cosmos.Json
 
             private static int? GetSystemStringIdLength17(ReadOnlySpan<byte> buffer)
             {
-                if (buffer.SequenceEqual(TokenBuffers.String30.Span))
+                if (buffer.SequenceEqual(TokenBuffers.String30.Span.Span))
                 {
                     return 30;
                 }
@@ -345,12 +345,12 @@ namespace Microsoft.Azure.Cosmos.Json
         /// <summary>
         /// Gets the SystemStringId for a particular system string.
         /// </summary>
-        /// <param name="utf8String">The system string to get the enum id for.</param>
+        /// <param name="utf8Span">The system string to get the enum id for.</param>
         /// <param name="systemStringId">The id of the system string if found.</param>
         /// <returns>The SystemStringId for a particular system string.</returns>
-        public static bool TryGetSystemStringId(ReadOnlySpan<byte> utf8String, out int systemStringId)
+        public static bool TryGetSystemStringId(Utf8Span utf8Span, out int systemStringId)
         {
-            int? id = SystemStrings.GetSystemStringId(utf8String);
+            int? id = SystemStrings.GetSystemStringId(utf8Span);
             if (!id.HasValue)
             {
                 systemStringId = default;
@@ -361,27 +361,15 @@ namespace Microsoft.Azure.Cosmos.Json
             return true;
         }
 
-        public static bool TryGetSystemStringById(int id, out string systemString)
+        public static bool TryGetSystemStringById(int id, out UtfAllString systemString)
         {
-            if (id >= SystemStrings.Utf16Values.Length)
+            if (id >= SystemStrings.UtfAllStringValues.Length)
             {
                 systemString = default;
                 return false;
             }
 
-            systemString = SystemStrings.Utf16Values[id];
-            return true;
-        }
-
-        public static bool TryGetUtf8SystemStringById(int id, out ReadOnlyMemory<byte> utf8SystemString)
-        {
-            if (id >= SystemStrings.Utf8Values.Length)
-            {
-                utf8SystemString = default;
-                return false;
-            }
-
-            utf8SystemString = SystemStrings.Utf8Values[id];
+            systemString = SystemStrings.UtfAllStringValues[id];
             return true;
         }
     }
