+++ Microsoft.Azure.Cosmos/src/Json/JsonNavigator.JsonBinaryNavigator.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Azure.Cosmos.Json
     using System.IO;
     using System.Runtime.InteropServices;
     using System.Text;
+    using Microsoft.Azure.Cosmos.Core.Utf8;
 
     /// <summary>
     /// Partial class that wraps the private JsonTextNavigator
@@ -107,15 +108,18 @@ namespace Microsoft.Azure.Cosmos.Json
             }
 
             /// <inheritdoc />
-            public override bool TryGetBufferedUtf8StringValue(
+            public override bool TryGetBufferedStringValue(
                 IJsonNavigatorNode stringNode,
-                out ReadOnlyMemory<byte> bufferedUtf8StringValue)
+                out Utf8Memory bufferedUtf8StringValue)
             {
                 ReadOnlyMemory<byte> buffer = JsonBinaryNavigator.GetNodeOfType(
                     JsonNodeType.String,
                     stringNode);
 
-                return JsonBinaryEncoding.TryGetBufferedUtf8StringValue(buffer, this.jsonStringDictionary, out bufferedUtf8StringValue);
+                return JsonBinaryEncoding.TryGetBufferedStringValue(
+                    Utf8Memory.UnsafeCreateNoValidation(buffer),
+                    this.jsonStringDictionary,
+                    out bufferedUtf8StringValue);
             }
 
             /// <inheritdoc />
@@ -124,7 +128,9 @@ namespace Microsoft.Azure.Cosmos.Json
                 ReadOnlyMemory<byte> buffer = JsonBinaryNavigator.GetNodeOfType(
                     JsonNodeType.String,
                     stringNode);
-                return JsonBinaryEncoding.GetStringValue(buffer, this.jsonStringDictionary);
+                return JsonBinaryEncoding.GetStringValue(
+                    Utf8Memory.UnsafeCreateNoValidation(buffer),
+                    this.jsonStringDictionary);
             }
 
             /// <inheritdoc />
@@ -398,17 +404,17 @@ namespace Microsoft.Azure.Cosmos.Json
                 string propertyName,
                 out ObjectProperty objectProperty)
             {
-                ReadOnlyMemory<byte> buffer = JsonBinaryNavigator.GetNodeOfType(
+                _ = JsonBinaryNavigator.GetNodeOfType(
                     JsonNodeType.Object,
                     objectNode);
 
-                ReadOnlySpan<byte> utf8StringPropertyName = Encoding.UTF8.GetBytes(propertyName);
+                Utf8Span utf8StringPropertyName = Utf8Span.TranscodeUtf16(propertyName);
                 foreach (ObjectProperty objectPropertyNode in this.GetObjectProperties(objectNode))
                 {
-                    if (this.TryGetBufferedUtf8StringValue(objectPropertyNode.NameNode, out ReadOnlyMemory<byte> bufferedUtf8StringValue))
+                    if (this.TryGetBufferedStringValue(objectPropertyNode.NameNode, out Utf8Memory bufferedUtf8StringValue))
                     {
                         // First try and see if we can avoid materializing the UTF16 string.
-                        if (utf8StringPropertyName.SequenceEqual(bufferedUtf8StringValue.Span))
+                        if (utf8StringPropertyName.Equals(bufferedUtf8StringValue.Span))
                         {
                             objectProperty = objectPropertyNode;
                             return true;
@@ -424,7 +430,7 @@ namespace Microsoft.Azure.Cosmos.Json
                     }
                 }
 
-                objectProperty = default(ObjectProperty);
+                objectProperty = default;
                 return false;
             }
