+++ Microsoft.Azure.Cosmos/src/Json/JsonStringDictionary.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Azure.Cosmos.Json
     using System.Collections.Generic;
     using System.Text;
     using Microsoft.Azure.Cosmos.Core.Collections;
+    using Microsoft.Azure.Cosmos.Core.Utf8;
 
 #if INTERNAL
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
@@ -20,8 +21,7 @@ namespace Microsoft.Azure.Cosmos.Json
     {
         private const int MaxStackAllocSize = 4 * 1024;
 
-        private readonly ReadOnlyMemory<byte>[] utf8Strings;
-        private readonly string[] utf16Strings;
+        private readonly UtfAllString[] strings;
         private readonly Trie<byte, int> utf8StringToIndex;
 
         private int size;
@@ -33,8 +33,7 @@ namespace Microsoft.Azure.Cosmos.Json
                 throw new ArgumentOutOfRangeException($"{nameof(capacity)} must be a non negative integer.");
             }
 
-            this.utf8Strings = new ReadOnlyMemory<byte>[capacity];
-            this.utf16Strings = new string[capacity];
+            this.strings = new UtfAllString[capacity];
             this.utf8StringToIndex = new Trie<byte, int>(capacity);
         }
 
@@ -44,34 +43,33 @@ namespace Microsoft.Azure.Cosmos.Json
             Span<byte> utfString = utf8Length < JsonStringDictionary.MaxStackAllocSize ? stackalloc byte[utf8Length] : new byte[utf8Length];
             Encoding.UTF8.GetBytes(value, utfString);
 
-            return this.TryAddString(utfString, out index);
+            return this.TryAddString(Utf8Span.UnsafeFromUtf8BytesNoValidation(utfString), out index);
         }
 
-        public bool TryAddString(ReadOnlySpan<byte> value, out int index)
+        public bool TryAddString(Utf8Span value, out int index)
         {
             // If the string already exists, then just return that index.
-            if (this.utf8StringToIndex.TryGetValue(value, out index))
+            if (this.utf8StringToIndex.TryGetValue(value.Span, out index))
             {
                 return true;
             }
 
             // If we are at capacity just return false.
-            if (this.size == this.utf8Strings.Length)
+            if (this.size == this.strings.Length)
             {
                 index = default;
                 return false;
             }
 
             index = this.size;
-            this.utf8Strings[this.size] = value.ToArray();
-            this.utf16Strings[this.size] = Encoding.UTF8.GetString(value);
-            this.utf8StringToIndex.AddOrUpdate(value, index);
+            this.strings[this.size] = UtfAllString.Create(value.ToString());
+            this.utf8StringToIndex.AddOrUpdate(value.Span, index);
             this.size++;
 
             return true;
         }
 
-        public bool TryGetStringAtIndex(int index, out string value)
+        public bool TryGetStringAtIndex(int index, out UtfAllString value)
         {
             if ((index < 0) || (index >= this.size))
             {
@@ -79,19 +77,7 @@ namespace Microsoft.Azure.Cosmos.Json
                 return false;
             }
 
-            value = this.utf16Strings[index];
-            return true;
-        }
-
-        public bool TryGetUtf8StringAtIndex(int index, out ReadOnlyMemory<byte> value)
-        {
-            if ((index < 0) || (index >= this.size))
-            {
-                value = default;
-                return false;
-            }
-
-            value = this.utf8Strings[index];
+            value = this.strings[index];
             return true;
         }
 
@@ -106,7 +92,7 @@ namespace Microsoft.Azure.Cosmos.Json
             for (int i = 0; i < userStrings.Count; i++)
             {
                 string userString = userStrings[i];
-                if (!jsonStringDictionary.TryAddString(Encoding.UTF8.GetBytes(userString).AsSpan(), out int index))
+                if (!jsonStringDictionary.TryAddString(Utf8Span.TranscodeUtf16(userString), out int index))
                 {
                     throw new ArgumentException($"Failed to add {userString} to {nameof(JsonStringDictionary)}.");
                 }
