+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/JsonWriterTests.cs
@@ -8,6 +8,7 @@
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using System.Reflection;
     using System.Globalization;
+    using Microsoft.Azure.Cosmos.Core.Utf8;
 
     [TestClass]
     public class JsonWriterTests
@@ -401,7 +402,7 @@
         public void SystemStringTest()
         {
             int systemStringId = 0;
-            while (JsonBinaryEncoding.TryGetSystemStringById(systemStringId, out string systemString))
+            while (JsonBinaryEncoding.TryGetSystemStringById(systemStringId, out UtfAllString systemString))
             {
                 string expectedString = "\"" + systemString + "\"";
                 // remove formatting on the json and also replace "/" with "\/" since newtonsoft is dumb.
@@ -417,7 +418,7 @@
 
                 JsonToken[] tokensToWrite =
                 {
-                    JsonToken.String(systemString)
+                    JsonToken.String(systemString.Utf16String)
                 };
 
                 this.VerifyWriter(tokensToWrite, expectedString);
@@ -488,8 +489,8 @@
             for (int i = 0; i < OneByteCount + 1; i++)
             {
                 string userEncodedString = "a" + i.ToString();
-                Assert.IsTrue(jsonStringDictionary.TryGetStringAtIndex(i, out string value));
-                Assert.AreEqual(userEncodedString, value);
+                Assert.IsTrue(jsonStringDictionary.TryGetStringAtIndex(i, out UtfAllString value));
+                Assert.AreEqual(userEncodedString, value.Utf16String);
             }
         }
         #endregion
@@ -1736,7 +1737,7 @@
                             string stringValue = (token as JsonStringToken).Value;
                             if (writeAsUtf8String)
                             {
-                                jsonWriter.WriteStringValue(Encoding.UTF8.GetBytes(stringValue));
+                                jsonWriter.WriteStringValue(Utf8Span.TranscodeUtf16(stringValue));
                             }
                             else
                             {
@@ -1765,7 +1766,7 @@
                             string fieldNameValue = (token as JsonFieldNameToken).Value;
                             if (writeAsUtf8String)
                             {
-                                jsonWriter.WriteFieldName(Encoding.UTF8.GetBytes(fieldNameValue));
+                                jsonWriter.WriteFieldName(Utf8Span.TranscodeUtf16(fieldNameValue));
                             }
                             else
                             {
